Delete Note Success

+++ src/pages/Home/Home.jsx
@@ -1,8 +1,10 @@
-import Navbar from '../../layout/Navbar'
+import Swal from 'sweetalert2'
 import axios from 'axios'
-import { useEffect, useState } from 'react'
+import Navbar from '../../layout/Navbar'
 import SearchBar from '../../components/searchBar'
 import { Icon } from '@iconify/react'
+import { useEffect, useState } from 'react'
+
 const Home = () => {
   const [notes, setNotes] = useState([])
 
@@ -10,6 +12,7 @@ const Home = () => {
     getNote()
   }, [])
 
+  // Get Note
   const getNote = async () => {
     try {
       const token = localStorage.getItem('token')
@@ -25,17 +28,35 @@ const Home = () => {
     }
   }
 
+  // Delete Note
   const deleteNote = async (id_note) => {
-    try {
-      const token = localStorage.getItem('token')
-      await axios.delete(`http://localhost:3000/notes/${id_note}`, {
-        headers: {
-          Authorization: token
-        }
-      })
-      getNote()
-    } catch (error) {
-      console.log(error)
+    const result = await Swal.fire({
+      title: 'Apakah anda yakin?',
+      text: 'Apakah anda yakin ingin menghapus note ini?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Hapus!'
+    })
+
+    if (result.isConfirmed) {
+      try {
+        const token = localStorage.getItem('token')
+        await axios.delete(`http://localhost:3000/notes/${id_note}`, {
+          headers: {
+            Authorization: token
+          }
+        })
+
+        setNotes((prevNotes) => {
+          return prevNotes.filter((note) => note.id_note !== id_note)
+        })
+        Swal.fire('Dihapus!', 'Note anda berhasil dihapus.', 'success')
+      } catch (error) {
+        console.log(error)
+        Swal.fire('Gagal!', 'Note anda gagal dihapus.', 'error')
+      }
     }
   }
 
@@ -54,7 +75,7 @@ const Home = () => {
             <a
               key={note.id_note}
               href="#"
-              className="flex justify-between items-center max-w p-6 border  rounded-lg shadow  bg-gray-800 border-gray-700 hover:bg-gray-700 mb-3"
+              className="flex justify-between items-center max-w p-6 border  rounded-lg shadow  bg-gray-800 border-gray-700 mb-3"
             >
               <div className="block">
                 <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-800 dark:text-white">
@@ -62,7 +83,10 @@ const Home = () => {
                 </h5>
                 <p className="font-normal text-gray-700 dark:text-gray-400">{note.date}</p>
               </div>
-              <button className="py-4 px-5 bg-red-500 rounded-md" onClick={deleteNote}>
+              <button
+                className="py-4 px-5 bg-red-500 rounded-md hover:bg-red-400"
+                onClick={() => deleteNote(note.id_note)}
+              >
                 <Icon icon="material-symbols:delete" className="text-2xl text-white" />
               </button>
             </a>
